refactor: unify FW authenticity check selectors
