+++ packages/suite/src/selectors/suite/suiteAuthenticityChecksSelectors.ts
@@ -1,36 +1,24 @@
 import {
-    hashCheckErrorScenarios,
+    isHardHashCheckError,
+    isHardRevisionCheckError,
     isSkippedHashCheckError,
     isSkippedRevisionCheckError,
-    revisionCheckErrorScenarios,
 } from '@suite-common/firmware-authenticity';
 import { Feature, selectIsFeatureDisabled } from '@suite-common/message-system';
-import { isDeviceAcquired } from '@suite-common/suite-utils';
-import { selectIsEntropyCheckFailed, selectSelectedDevice } from '@suite-common/wallet-core';
+import {
+    selectFirmwareHashCheckError,
+    selectFirmwareRevisionCheckError,
+    selectIsEntropyCheckFailed,
+} from '@suite-common/wallet-core';
 
 import { AppState } from 'src/types/suite';
 
 import { selectIsEntropyCheckEnabled } from './suiteSelectors';
 
-/**
- * Get firmware revision check error, or null if check was successful / skipped.
- */
-export const selectFirmwareRevisionCheckError = (state: AppState) => {
-    const device = selectSelectedDevice(state);
-    if (!isDeviceAcquired(device) || !device.authenticityChecks) return null;
-    const checkResult = device.authenticityChecks.firmwareRevision;
-
-    // null means not performed, then don't consider it failed
-    if (!checkResult || checkResult.success) return null;
-
-    if (isSkippedRevisionCheckError(checkResult.error)) return null;
-
-    return checkResult.error;
-};
-
 export const selectFirmwareRevisionCheckErrorIfEnabled = (state: AppState) => {
     const revisionCheckError = selectFirmwareRevisionCheckError(state);
     if (revisionCheckError === null) return null;
+    if (isSkippedRevisionCheckError(revisionCheckError)) return null;
 
     const isFirmwareRevisionCheckDisabled =
         !state.suite.settings.enabledSecurityChecks.firmwareRevision;
@@ -42,25 +30,10 @@ export const selectFirmwareRevisionCheckErrorIfEnabled = (state: AppState) => {
     return revisionCheckError;
 };
 
-/**
- * Get firmware hash check error, or null if check was successful / skipped.
- */
-export const selectFirmwareHashCheckError = (state: AppState) => {
-    const device = selectSelectedDevice(state);
-    if (!isDeviceAcquired(device) || !device.authenticityChecks) return null;
-    const checkResult = device.authenticityChecks.firmwareHash;
-
-    // null means not performed, then don't consider it failed
-    if (!checkResult || checkResult.success) return null;
-
-    if (isSkippedHashCheckError(checkResult.error)) return null;
-
-    return checkResult.error;
-};
-
 export const selectFirmwareHashCheckErrorIfEnabled = (state: AppState) => {
     const hashCheckError = selectFirmwareHashCheckError(state);
     if (hashCheckError === null) return null;
+    if (isSkippedHashCheckError(hashCheckError)) return null;
 
     const isFirmwareHashCheckDisabled = !state.suite.settings.enabledSecurityChecks.firmwareHash;
     if (isFirmwareHashCheckDisabled) return null;
@@ -90,14 +63,9 @@ export const selectFirmwareHashCheckErrorIfEnabled = (state: AppState) => {
  */
 export const selectIsFirmwareAuthenticityCheckEnabledAndHardFailed = (state: AppState) => {
     const revisionError = selectFirmwareRevisionCheckErrorIfEnabled(state);
-    const isRevisionHardError =
-        revisionError !== null && revisionCheckErrorScenarios[revisionError].type === 'hardModal';
-
     const hashError = selectFirmwareHashCheckErrorIfEnabled(state);
-    const isHashHardError =
-        hashError !== null && hashCheckErrorScenarios[hashError].type === 'hardModal';
 
-    return isRevisionHardError || isHashHardError;
+    return isHardRevisionCheckError(revisionError) || isHardHashCheckError(hashError);
 };
 
 /**
