+++ suite-common/wallet-core/src/device/deviceSelectors.ts
@@ -515,3 +515,31 @@ export const selectDeviceDefaultBackupType = createMemoizedSelector(
 export const selectStandardWalletDevice = createMemoizedSelector([selectDevices], devices =>
     devices.find(device => device.useEmptyPassphrase),
 );
+
+/**
+ * Get firmware revision check error, or null if check was successful / skipped.
+ */
+export const selectFirmwareRevisionCheckError = (state: DeviceRootState) => {
+    const device = selectSelectedDevice(state);
+    if (!deviceUtils.isDeviceAcquired(device) || !device.authenticityChecks) return null;
+    const checkResult = device.authenticityChecks.firmwareRevision;
+
+    // null means not performed, then don't consider it failed
+    if (!checkResult || checkResult.success) return null;
+
+    return checkResult.error;
+};
+
+/**
+ * Get firmware hash check error, or null if check was successful / skipped.
+ */
+export const selectFirmwareHashCheckError = (state: DeviceRootState) => {
+    const device = selectSelectedDevice(state);
+    if (!deviceUtils.isDeviceAcquired(device) || !device.authenticityChecks) return null;
+    const checkResult = device.authenticityChecks.firmwareHash;
+
+    // null means not performed, then don't consider it failed
+    if (!checkResult || checkResult.success) return null;
+
+    return checkResult.error;
+};
